#638 moved all xsd casts to xsd package; reuse CastFunction class
