+++ core/queryalgebra/evaluation/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/function/IntegerCast.java
@@ -7,17 +7,6 @@
  *******************************************************************************/
 package org.eclipse.rdf4j.query.algebra.evaluation.function;
 
-import java.math.BigInteger;
-
-import org.eclipse.rdf4j.model.IRI;
-import org.eclipse.rdf4j.model.Literal;
-import org.eclipse.rdf4j.model.Value;
-import org.eclipse.rdf4j.model.ValueFactory;
-import org.eclipse.rdf4j.model.datatypes.XMLDatatypeUtil;
-import org.eclipse.rdf4j.model.vocabulary.XMLSchema;
-import org.eclipse.rdf4j.query.algebra.evaluation.ValueExprEvaluationException;
-import org.eclipse.rdf4j.query.algebra.evaluation.util.QueryEvaluationUtil;
-
 /**
  * A {@link Function} that tries to cast its argument to an <tt>xsd:integer</tt> .
  * 
@@ -25,66 +14,6 @@ import org.eclipse.rdf4j.query.algebra.evaluation.util.QueryEvaluationUtil;
  * @deprecated use {@link org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.IntegerCast} instead
  */
 @Deprecated
-public class IntegerCast implements Function {
+public class IntegerCast extends org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.IntegerCast {
 
-	public String getURI() {
-		return XMLSchema.INTEGER.toString();
-	}
-
-	public Literal evaluate(ValueFactory valueFactory, Value... args)
-		throws ValueExprEvaluationException
-	{
-		if (args.length != 1) {
-			throw new ValueExprEvaluationException(
-					"xsd:integer cast requires exactly 1 argument, got " + args.length);
-		}
-
-		if (args[0] instanceof Literal) {
-			Literal literal = (Literal)args[0];
-			IRI datatype = literal.getDatatype();
-
-			if (QueryEvaluationUtil.isStringLiteral(literal)) {
-				String integerValue = XMLDatatypeUtil.collapseWhiteSpace(literal.getLabel());
-				if (XMLDatatypeUtil.isValidInteger(integerValue)) {
-					return valueFactory.createLiteral(integerValue, XMLSchema.INTEGER);
-				}
-			}
-			else if (datatype != null) {
-				if (datatype.equals(XMLSchema.INTEGER)) {
-					return literal;
-				}
-				else if (XMLDatatypeUtil.isNumericDatatype(datatype)) {
-					// decimals, floats and doubles must be processed
-					// separately, see
-					// http://www.w3.org/TR/xpath-functions/#casting-from-primitive-to-primitive
-					BigInteger integerValue = null;
-					if (XMLSchema.DECIMAL.equals(datatype)
-							|| XMLDatatypeUtil.isFloatingPointDatatype(datatype))
-					{
-						integerValue = literal.decimalValue().toBigInteger();
-					}
-					else {
-						integerValue = literal.integerValue();
-					}
-					try {
-						return valueFactory.createLiteral(integerValue.toString(), XMLSchema.INTEGER);
-					}
-					catch (NumberFormatException e) {
-						throw new ValueExprEvaluationException(e.getMessage(), e);
-					}
-				}
-				else if (datatype.equals(XMLSchema.BOOLEAN)) {
-					try {
-						return valueFactory.createLiteral(literal.booleanValue() ? "1" : "0",
-								XMLSchema.INTEGER);
-					}
-					catch (IllegalArgumentException e) {
-						throw new ValueExprEvaluationException(e.getMessage(), e);
-					}
-				}
-			}
-		}
-
-		throw new ValueExprEvaluationException("Invalid argument for xsd:integer cast: " + args[0]);
-	}
 }
