+++ core/queryalgebra/evaluation/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/function/xsd/UnsignedIntCast.java
@@ -17,14 +17,14 @@ import org.eclipse.rdf4j.model.datatypes.XMLDatatypeUtil;
 import org.eclipse.rdf4j.model.vocabulary.XMLSchema;
 
 /**
- * A {@link IntegerDatatypeCast} that tries to cast its argument to an <tt>xsd:unsignedInt</tt> .
+ * A {@link IntegerCastFunction} that tries to cast its argument to an <tt>xsd:unsignedInt</tt> .
  * 
  * @author Jeen Broekstra
  */
-public class UnsignedIntCast extends IntegerDatatypeCast {
+public class UnsignedIntCast extends IntegerCastFunction {
 
 	@Override
-	protected IRI getIntegerDatatype() {
+	protected IRI getXsdDatatype() {
 		return XMLSchema.UNSIGNED_INT;
 	}
