+++ core/queryalgebra/evaluation/src/main/resources/META-INF/services/org.eclipse.rdf4j.query.algebra.evaluation.function.Function
@@ -1,9 +1,4 @@
-org.eclipse.rdf4j.query.algebra.evaluation.function.StringCast
-org.eclipse.rdf4j.query.algebra.evaluation.function.FloatCast
-org.eclipse.rdf4j.query.algebra.evaluation.function.DoubleCast
-org.eclipse.rdf4j.query.algebra.evaluation.function.DecimalCast
-org.eclipse.rdf4j.query.algebra.evaluation.function.BooleanCast
-org.eclipse.rdf4j.query.algebra.evaluation.function.DateTimeCast
+
 org.eclipse.rdf4j.query.algebra.evaluation.function.datetime.Day
 org.eclipse.rdf4j.query.algebra.evaluation.function.datetime.Hours
 org.eclipse.rdf4j.query.algebra.evaluation.function.datetime.Minutes
@@ -39,15 +34,21 @@ org.eclipse.rdf4j.query.algebra.evaluation.function.string.StrLen
 org.eclipse.rdf4j.query.algebra.evaluation.function.string.StrStarts
 org.eclipse.rdf4j.query.algebra.evaluation.function.string.Substring
 org.eclipse.rdf4j.query.algebra.evaluation.function.string.UpperCase
+org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.BooleanCast
 org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.ByteCast
+org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.DateTimeCast
+org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.DecimalCast
+org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.DoubleCast
+org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.FloatCast
 org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.IntCast
-org.eclipse.rdf4j.query.algebra.evaluation.function.IntegerCast
+org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.IntegerCast
 org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.LongCast
 org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.NegativeIntegerCast
 org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.NonNegativeIntegerCast
 org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.NonPositiveIntegerCast
 org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.PositiveIntegerCast
 org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.ShortCast
+org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.StringCast
 org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.UnsignedByteCast
 org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.UnsignedIntCast
 org.eclipse.rdf4j.query.algebra.evaluation.function.xsd.UnsignedLongCast
