+++ core/queryalgebra/evaluation/src/test/java/org/eclipse/rdf4j/query/algebra/evaluation/function/xsd/TestStringCast.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/org/documents/edl-v10.php.
  *******************************************************************************/
-package org.eclipse.rdf4j.query.algebra.evaluation.function;
+package org.eclipse.rdf4j.query.algebra.evaluation.function.xsd;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -25,9 +25,9 @@ import org.junit.Test;
 /**
  * @author jeen
  */
-public class TestDateTimeCast {
+public class TestStringCast {
 
-	private DateTimeCast dtCast;
+	private StringCast stringCast;
 
 	private ValueFactory f = SimpleValueFactory.getInstance();
 
@@ -38,7 +38,7 @@ public class TestDateTimeCast {
 	public void setUp()
 		throws Exception
 	{
-		dtCast = new DateTimeCast();
+		stringCast = new StringCast();
 	}
 
 	/**
@@ -52,11 +52,11 @@ public class TestDateTimeCast {
 
 	@Test
 	public void testCastPlainLiteral() {
-		Literal plainLit = f.createLiteral("1999-09-09T00:00:01");
+		Literal plainLit = f.createLiteral("foo");
 		try {
-			Literal result = dtCast.evaluate(f, plainLit);
+			Literal result = stringCast.evaluate(f, plainLit);
 			assertNotNull(result);
-			assertEquals(XMLSchema.DATETIME, result.getDatatype());
+			assertEquals(XMLSchema.STRING, result.getDatatype());
 		}
 		catch (ValueExprEvaluationException e) {
 			fail(e.getMessage());
@@ -64,13 +64,26 @@ public class TestDateTimeCast {
 	}
 
 	@Test
-	public void testCastDateLiteral() {
-		Literal dateLit = f.createLiteral("1999-09-09", XMLSchema.DATE);
+	public void testCastLangtagLiteral() {
+		Literal langLit = f.createLiteral("foo", "en");
 		try {
-			Literal result = dtCast.evaluate(f, dateLit);
-			assertNotNull(result);
-			assertEquals(XMLSchema.DATETIME, result.getDatatype());
+			Literal result = stringCast.evaluate(f, langLit);
+			fail("casting of language-tagged literal to xsd:string should result in type error");
+		}
+		catch (ValueExprEvaluationException e) {
+			// do nothing, expected
+		}
+	}
 
+	@Test
+	public void testCastIntegerLiteral() {
+		Literal intLit = f.createLiteral(10);
+		try {
+			Literal result = stringCast.evaluate(f, intLit);
+			assertNotNull(result);
+			assertEquals(XMLSchema.STRING, result.getDatatype());
+			assertFalse(result.getLanguage().isPresent());
+			assertEquals("10", result.getLabel());
 		}
 		catch (ValueExprEvaluationException e) {
 			fail(e.getMessage());
@@ -82,9 +95,9 @@ public class TestDateTimeCast {
 		String lexVal = "2000-01-01T00:00:00";
 		Literal dtLit = f.createLiteral(XMLDatatypeUtil.parseCalendar(lexVal));
 		try {
-			Literal result = dtCast.evaluate(f, dtLit);
+			Literal result = stringCast.evaluate(f, dtLit);
 			assertNotNull(result);
-			assertEquals(XMLSchema.DATETIME, result.getDatatype());
+			assertEquals(XMLSchema.STRING, result.getDatatype());
 			assertFalse(result.getLanguage().isPresent());
 			assertEquals(lexVal, result.getLabel());
 		}
@@ -93,4 +106,19 @@ public class TestDateTimeCast {
 		}
 	}
 
+	@Test
+	public void testCastUnknownDatatypedLiteral() {
+		String lexVal = "foobar";
+		Literal dtLit = f.createLiteral(lexVal, f.createIRI("foo:unknownDt"));
+		try {
+			Literal result = stringCast.evaluate(f, dtLit);
+			assertNotNull(result);
+			assertEquals(XMLSchema.STRING, result.getDatatype());
+			assertFalse(result.getLanguage().isPresent());
+			assertEquals(lexVal, result.getLabel());
+		}
+		catch (ValueExprEvaluationException e) {
+			fail(e.getMessage());
+		}
+	}
 }
